user can delete authorized thread

+++ app/Http/Controllers/ThreadsController.php
@@ -112,7 +112,13 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-//        $thread->replies()->delete();
+        $this->authorize('update', $thread);
+        if ($thread->user_id != auth()->id()) {
+            if (request()->wantsJson()) {
+                abort(403, 'You do not have permissions to do this ');
+            }
+            return redirect('/login');
+        }
         $thread->delete();
         if (request()->wantsJson()) {
             return response([], 204);
